Added perlin configs, updated generator

+++ pkg/terrains/models/TerrainChunk.go
@@ -20,10 +20,10 @@ type TerrainChunk struct {
 	ResolutionPerPatch  int                   `gorm:"type:int;not null" json:"resolutionPerPatch"`
 	HeightmapResolution int                   `gorm:"type:int;not null" json:"heightmapResolution"`
 	AlphamapResolution  int                   `gorm:"type:int;not null" json:"alphamapResolution"`
-	Seed                int64                 `gorm:"type:int;not null" json:"seed"`
+	Seed                int                   `gorm:"type:int;not null" json:"seed"`
 	Frequency           float32               `gorm:"type:float;not null" json:"frequency"`
 	Gain                float32               `gorm:"type:float;not null" json:"gain"`
-	Octaves             float32               `gorm:"type:float;not null" json:"octaves"`
+	Octaves             int                   `gorm:"type:float;not null" json:"octaves"`
 	Lacunarity          float32               `gorm:"type:float;not null" json:"lacunarity"`
 	Created             time.Time             `gorm:"autoCreateTime" json:"created"`
 	Updated             time.Time             `gorm:"autoUpdateTime" json:"updated"`
